Added color palette
