+++ projects/lfx-component-lib/.storybook/main.ts
@@ -8,7 +8,8 @@ const config: StorybookConfig = {
     '@storybook/addon-essentials',
     '@chromatic-com/storybook',
     '@storybook/addon-interactions',
-    '@storybook/addon-designs'
+    '@storybook/addon-designs',
+    '@storybook/addon-docs',
   ],
   framework: {
     name: '@storybook/angular',
