+++ projects/lfx-component-lib/src/lib/themes/color-palette.mdx
@@ -1,0 +1,142 @@
+import { Meta, ColorPalette, ColorItem } from "@storybook/blocks";
+import colors from "../configs/color-palette.json";
+
+<Meta title="LFX Components/Themes/Color Palette" />
+
+# Color Palette
+
+Our design system's color palette helps maintain consistent styling across the application.
+
+<ColorPalette>
+  <ColorItem
+    title="bg-primary"
+    subtitle="Background Primary"
+    colors={{
+      WildWatermelon: colors.blue[500],
+    }}
+  />
+  <ColorItem
+    title="bg-secondary"
+    subtitle="Background Secondary"
+    colors={{
+      Blue: colors.blue[50],
+    }}
+  />
+  <ColorItem
+    title="bg-message"
+    subtitle="Background Message"
+    colors={{
+      Notice: colors.blue[50],
+      Warning: colors.orange[50],
+      Success: colors.green[50],
+      Error: colors.red[50],
+    }}
+  />
+
+<ColorItem
+  title="bg-grey"
+  subtitle="Background Grey"
+  colors={{
+    50: colors.grey[50],
+    100: colors.grey[100],
+    200: colors.grey[200],
+    300: colors.grey[300],
+    400: colors.grey[400],
+    500: colors.grey[500],
+    600: colors.grey[600],
+    700: colors.grey[700],
+    800: colors.grey[800],
+    900: colors.grey[900],
+    1000: colors.grey[1000],
+  }}
+/>
+
+<ColorItem
+  title="bg-blue"
+  subtitle="Background Blue"
+  colors={{
+    50: colors.blue[50],
+    100: colors.blue[100],
+    200: colors.blue[200],
+    300: colors.blue[300],
+    400: colors.blue[400],
+    500: colors.blue[500],
+    600: colors.blue[600],
+    700: colors.blue[700],
+    800: colors.blue[800],
+    900: colors.blue[900],
+    1000: colors.blue[1000],
+  }}
+/>
+
+<ColorItem
+  title="bg-green"
+  subtitle="Background Green"
+  colors={{
+    50: colors.green[50],
+    100: colors.green[100],
+    200: colors.green[200],
+    300: colors.green[300],
+    400: colors.green[400],
+    500: colors.green[500],
+    600: colors.green[600],
+    700: colors.green[700],
+    800: colors.green[800],
+    900: colors.green[900],
+    1000: colors.green[1000],
+  }}
+/>
+
+<ColorItem
+  title="bg-red"
+  subtitle="Background Red"
+  colors={{
+    50: colors.red[50],
+    100: colors.red[100],
+    200: colors.red[200],
+    300: colors.red[300],
+    400: colors.red[400],
+    500: colors.red[500],
+    600: colors.red[600],
+    700: colors.red[700],
+    800: colors.red[800],
+    900: colors.red[900],
+  }}
+/>
+
+<ColorItem
+  title="bg-yellow"
+  subtitle="Background Yellow"
+  colors={{
+    50: colors.yellow[50],
+    100: colors.yellow[100],
+    200: colors.yellow[200],
+    300: colors.yellow[300],
+    400: colors.yellow[400],
+    500: colors.yellow[500],
+    600: colors.yellow[600],
+    700: colors.yellow[700],
+    800: colors.yellow[800],
+    900: colors.yellow[900],
+    1000: colors.yellow[1000],
+  }}
+/>
+
+  <ColorItem
+    title="bg-orange"
+    subtitle="Background Orange"
+    colors={{
+      50: colors.orange[50],
+      100: colors.orange[100],
+      200: colors.orange[200],
+      300: colors.orange[300],
+      400: colors.orange[400],
+      500: colors.orange[500],
+      600: colors.orange[600],
+      700: colors.orange[700],
+      800: colors.orange[800],
+      900: colors.orange[900],
+      1000: colors.orange[1000],
+    }}
+  />
+</ColorPalette>
